Re: 어드민 로직 분리
