+++ src/main/java/org/example/hanmo/controller/UserController.java
@@ -77,18 +77,4 @@ public class UserController {
     userService.logout(tempToken);
     return ResponseEntity.ok("로그아웃 되었습니다.");
   }
-
-  @Operation(summary = "관리자 로그인 (테스트용입니다. jwt나올 시 변경, 지금은 임시토큰)")
-  @PostMapping("/login/admin")
-  public ResponseEntity<String> loginAdmin(@RequestBody AdminRequestDto request) {
-    String tempToken = userService.loginAdmin(request);
-    return ResponseEntity.ok().header("tempToken", tempToken).body("관리자 로그인 되었습니다.");
-  }
-
-  @Operation(summary = "관리자 추가 정보 입력")
-  @PutMapping("/signup/admin")
-  public ResponseEntity<String> addAdminInfo(@RequestBody AdminRequestDto dto) {
-    userService.addAdminInfo(dto);
-    return ResponseEntity.ok("추가 정보가 입력되었습니다.");
-  }
 }
