+++ src/main/java/org/example/hanmo/service/AdminService.java
@@ -1,0 +1,9 @@
+package org.example.hanmo.service;
+
+import org.example.hanmo.dto.user.request.AdminRequestDto;
+
+public interface AdminService {
+    String loginAdmin(AdminRequestDto requestDto);
+
+    void addAdminInfo(AdminRequestDto dto);
+}
