+++ src/main/java/org/example/hanmo/service/UserService.java
@@ -1,6 +1,5 @@
 package org.example.hanmo.service;
 
-import org.example.hanmo.dto.user.request.AdminRequestDto;
 import org.example.hanmo.dto.user.request.UserLoginRequestDto;
 import org.example.hanmo.dto.user.request.UserSignUpRequestDto;
 import org.example.hanmo.dto.user.response.UserProfileResponseDto;
@@ -14,10 +13,6 @@ public interface UserService {
   void withdrawUser(String phoneNumber);
 
   String loginUser(UserLoginRequestDto requestDto);
-
-  String loginAdmin(AdminRequestDto requestDto);
-
-  void addAdminInfo(AdminRequestDto dto);
 
   UserProfileResponseDto getUserProfile(String tempToken);
